+++ TorreWebapp/Models/Person/Project.cs
@@ -17,10 +17,10 @@ namespace TorreWebapp.Models
         public string toMonth { get; set; }
         public string toYear { get; set; }
         public bool highlighted { get; set; }
-        public int weight { get; set; }
+        public double weight { get; set; }
         public int verifications { get; set; }
         public int recommendations { get; set; }
         public List<object> media { get; set; }
-        public int rank { get; set; }
+        public double rank { get; set; }
     }
 }
